# Introduce in!(x, s::Set) to improve performance of unique()