This PR introduces `in!` function not to compute `ht_keyindex` twice. It is not exported yet, but the fallback would be:
``` julia
in!(x, s::AbstractSet) = in(x, s) ? true : (push!(s, x); false)
``` 

Master
``` julia
julia> using Random, BenchmarkTools

julia> @btime unique($([randstring(20) for i in 1:10]));
  483.954 ns (7 allocations: 848 bytes)

julia> @btime allunique($([randstring(20) for i in 1:10]));
  270.550 ns (4 allocations: 384 bytes)
```

PR
``` julia
julia> @btime unique($([randstring(20) for i in 1:10]));
  354.769 ns (7 allocations: 848 bytes)

julia> @btime allunique($([randstring(20) for i in 1:10]));
  270.236 ns (4 allocations: 384 bytes)
```

(`allunique` has been already optimized, but with `Dict`.)