@tkf 2022-05-03
Regarding `in!`, I think it'd be nice if we can keep it internal and explore we more generic and extensible abstract interface. FWIW, I personally prefer addressing #33758 and #45080 for solving this. #33758 (implemented in: [`PreludeDicts.modify!`](https://github.com/JuliaPreludes/PreludeDicts.jl#modify)) lets us define `in!`-like functions for arbitrary set types derived from dict types. #45080 lets us implement an interface like [`PreludeDicts.tryinsert!`](https://github.com/JuliaPreludes/PreludeDicts.jl#tryinsert) that is more generic; e.g., it can be used for (string) interning.