@petvana 2022-05-03
Initially, I was inspired by `get!` function for `Dict`, and I haven't found anything like this for `Set` to avoid hashing twice. There are two reasons why I selected `in!`. First, it is straightforward, short, and easy to understand. Secondly, `!in!` looks so imperative in the code. :sunglasses:
https://github.com/JuliaLang/julia/blob/709daeb613999993136e65657096ab2fc2eb5c3a/base/set.jl#L149-L151