@jakobnissen 2022-05-03
This also ties into the long-standing issue of providing token-based API for dictionaries (https://github.com/JuliaLang/julia/issues/24454), which would also expose an interface to solve this problem.
FWIW, this is the [default behaviour of Rust's `HashSet`](https://doc.rust-lang.org/std/collections/struct.HashSet.html#method.insert). Not that we should copy them, but it lends credence to it being a reasonable function to have.