@petvana 2022-05-03
Token-based API seems best in the long term. Here, I have two minor comments. It would be nice to store `age` from `Dict` in the token and check if the user is trying to shoot themself into a foot by using an outdated token. Secondly, the same token can be used for iterating over the collection to prevent the user to insert elements while iterating because it may be hard-to-debug (and is not well documented).

Meanwhile, the PR seems ready to be merged.