@tkf 2022-05-03
Functional API (`modify!`) is "better" than token-based API in the sense that the latter can be derived from the former while the former enables concurrent dictionary implementation. This is discussed in [Comparison to token-based API](https://github.com/JuliaPreludes/PreludeDicts.jl#comparison-to-token-based-api).